consolidate DeviceManager and IotHub
